--- components/MediaList/README.md
@@ -1,6 +1,0 @@
-MediaList
-========
-
-For lists that are entirely made of [`<Media />`](../Media) components, we should make an album or playlist, depending on the type of content.
-
-This component is expected to be used in conjunction with our Media component, and should be able to be used with that, OR with its more specialize counterparts, [`<Image />`](../MediaImage), [`<Video />`](../MediaImage), or [`<Audio />`](../MediaAudio).
