--- components/Tag/Tag.js
@@ -1,5 +1,0 @@
-import React from "react"
-
-export default function Tag(props) {
-  return <span className="tag h6 border rounded ml1">{props.children}</span>
-}
