+++ components/index.css
@@ -183,26 +183,6 @@ a:visited .link-post-title {
   max-height: 800px;
 }
 
-.tag {
-  padding: 2px;
-  padding-left: 4px;
-  padding-right: 4px;
-  margin-top: -2px;
-  display: inline-block;
-  text-transform: lowercase;
-  color: black;
-  cursor: pointer;
-  background-color: rgba(0,0,0,0.05);
-  border: 1px solid rgba(0,0,0,0.08);
-  opacity: 0.7;
-  user-select: none;
-}
-
-.tag a {
-  color: black;
-  text-decoration: none;
-}
-
 .flex {
   flex-grow: 3;
 }
@@ -251,7 +231,7 @@ a:visited .link-post-title {
 .article pre {
   background-color: rgb(44, 48, 58);
   border-radius: 6px;
-  color: rgba(256, 256, 256, 0.7);
+  color: rgba(256, 256, 256, 0.9);
   line-height: 1.6;
   overflow: scroll;
   padding: 2em;
@@ -260,7 +240,7 @@ a:visited .link-post-title {
 .article code {
   background-color: rgb(44, 48, 58);
   border-radius: 3px;
-  color: rgba(256, 256, 256, 0.7);
+  color: rgba(256, 256, 256, 0.9);
   padding: 0.2em 0.4em 0.2em 0.4em;
 }
