--- jestsetup.js
@@ -1,5 +1,0 @@
-import Enzyme from "enzyme"
-import Adapter from "enzyme-adapter-react-16"
-
-// React 16 Enzyme adapter
-Enzyme.configure({ adapter: new Adapter() })
