+++ package.json
@@ -12,9 +12,8 @@
     "postinstall": "npm run build",
     "start": "NODE_ENV=production node server.js -p $PORT",
     "dev": "node server.js -p $PORT",
-    "test": "npm-run-all -p test:*",
-    "test:lint": "tslint --project ./tsconfig.json",
-    "test:unit": "jest"
+    "test": "npm-run-all -p test:lint build",
+    "test:lint": "tslint --project ./tsconfig.json"
   },
   "dependencies": {
     "@babel/core": "^7.1.6",
@@ -34,7 +33,6 @@
     "firebase": "^5.6.0",
     "highlight.js": "^9.13.1",
     "isomorphic-unfetch": "^3.0.0",
-    "lodash": "^4.17.11",
     "marksy": "^6.1.0",
     "next": "^7.0.2",
     "npm-run-all": "^4.1.5",
@@ -47,20 +45,12 @@
     "xss-filters": "^1.2.7"
   },
   "devDependencies": {
-    "@types/enzyme": "^3.1.15",
-    "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/express": "^4.16.0",
-    "@types/jest": "^23.3.10",
-    "@types/lodash": "^4.14.118",
     "@types/next": "^7.0.5",
     "@types/nprogress": "0.0.29",
     "@types/react": "^16.7.11",
     "blog-o-matic": "^3.0.3",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "enzyme-to-json": "^3.3.4",
     "husky": "^1.2.0",
-    "jest": "^23.6.0",
     "tslint": "^5.11.0",
     "tslint-eslint-rules": "^5.4.0",
     "typescript": "^3.2.1"
@@ -89,32 +79,5 @@
       "pre-push": "npm-run-all test:lint build",
       "pre-publish": "npm-run-all test build"
     }
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "components/**/*.js",
-      "pages/**/*.js",
-      "utilities/**/*.js"
-    ],
-    "coveragePathIgnorePatterns": [
-      "node_modules",
-      "constants",
-      "vendor",
-      ".test.js",
-      ".d.ts$"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "jsx"
-    ],
-    "setupFiles": [
-      "<rootDir>/jestsetup.js"
-    ],
-    "snapshotSerializers": [
-      "enzyme-to-json/serializer"
-    ],
-    "testMatch": [
-      "**/*.test.(js|jsx)"
-    ]
   }
 }
