+++ pages/index.js
@@ -1,8 +1,8 @@
 import { get, isNumber } from "@civility/utilities"
 import Error from "next/error"
 import React from "react"
-import Layout from "../components/Layout/Layout"
-import Link from "../components/LinkPost/LinkPost"
+import Layout from "../components/Layout"
+import Link from "../components/LinkPost"
 import { readBlogMeta } from "../services/contentServices"
 import { shouldShowPost } from "../utilities/predicates"
 import { sortByDateString } from "../utilities/sorts"
